+++ packages/angular/src/charts.module.ts
@@ -11,6 +11,7 @@ import { GaugeChartComponent } from "./gauge-chart.component";
 import { LineChartComponent } from "./line-chart.component";
 import { PieChartComponent } from "./pie-chart.component";
 import { ScatterChartComponent } from "./scatter-chart.component";
+import { RadarChartComponent } from "./radar-chart.component";
 
 @NgModule({
 	imports: [
@@ -26,7 +27,8 @@ import { ScatterChartComponent } from "./scatter-chart.component";
 		GaugeChartComponent,
 		LineChartComponent,
 		PieChartComponent,
-		ScatterChartComponent
+		ScatterChartComponent,
+		RadarChartComponent
 	],
 	exports: [
 		BaseChart,
@@ -38,7 +40,8 @@ import { ScatterChartComponent } from "./scatter-chart.component";
 		GaugeChartComponent,
 		LineChartComponent,
 		PieChartComponent,
-		ScatterChartComponent
+		ScatterChartComponent,
+		RadarChartComponent
 	]
 })