+++ packages/core/demo/data/index.ts
@@ -105,8 +105,32 @@ let allDemoGroups = [
 				chartType: chartTypes.SimpleBarChart
 			},
 			{
+				options: barDemos.simpleBarEmptyStateOptions,
+				data: barDemos.simpleBarEmptyStateData,
+				chartType: chartTypes.SimpleBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.simpleBarSkeletonOptions,
+				data: barDemos.simpleBarSkeletonData,
+				chartType: chartTypes.SimpleBarChart,
+				isDemoExample: true
+			},
+			{
 				options: barDemos.groupedBarOptions,
 				data: barDemos.groupedBarData,
+				chartType: chartTypes.GroupedBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.groupedBarEmptyStateOptions,
+				data: barDemos.groupedBarEmptyStateData,
+				chartType: chartTypes.GroupedBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.groupedBarSkeletonOptions,
+				data: barDemos.groupedBarSkeletonData,
 				chartType: chartTypes.GroupedBarChart,
 				isDemoExample: true
 			},
@@ -119,6 +143,18 @@ let allDemoGroups = [
 			{
 				options: barDemos.stackedBarTimeSeriesOptions,
 				data: barDemos.stackedBarTimeSeriesData,
+				chartType: chartTypes.StackedBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.stackedBarEmptyStateOptions,
+				data: barDemos.stackedBarEmptyStateData,
+				chartType: chartTypes.StackedBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.stackedBarSkeletonOptions,
+				data: barDemos.stackedBarSkeletonData,
 				chartType: chartTypes.StackedBarChart,
 				isDemoExample: true
 			}
@@ -140,8 +176,32 @@ let allDemoGroups = [
 				isDemoExample: true
 			},
 			{
+				options: barDemos.simpleHorizontalBarEmptyStateOptions,
+				data: barDemos.simpleHorizontalBarEmptyStateData,
+				chartType: chartTypes.SimpleBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.simpleHorizontalBarSkeletonOptions,
+				data: barDemos.simpleHorizontalBarSkeletonData,
+				chartType: chartTypes.SimpleBarChart,
+				isDemoExample: true
+			},
+			{
 				options: barDemos.groupedHorizontalBarOptions,
 				data: barDemos.groupedHorizontalBarData,
+				chartType: chartTypes.GroupedBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.groupedHorizontalBarEmptyStateOptions,
+				data: barDemos.groupedHorizontalBarEmptyStateData,
+				chartType: chartTypes.GroupedBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.groupedHorizontalBarSkeletonOptions,
+				data: barDemos.groupedHorizontalBarSkeletonData,
 				chartType: chartTypes.GroupedBarChart,
 				isDemoExample: true
 			},
@@ -154,6 +214,18 @@ let allDemoGroups = [
 			{
 				options: barDemos.stackedHorizontalBarTimeSeriesOptions,
 				data: barDemos.stackedHorizontalBarTimeSeriesData,
+				chartType: chartTypes.StackedBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.stackedHorizontalBarEmptyStateOptions,
+				data: barDemos.stackedHorizontalBarEmptyStateData,
+				chartType: chartTypes.StackedBarChart,
+				isDemoExample: true
+			},
+			{
+				options: barDemos.stackedHorizontalBarSkeletonOptions,
+				data: barDemos.stackedHorizontalBarSkeletonData,
 				chartType: chartTypes.StackedBarChart,
 				isDemoExample: true
 			}
@@ -179,6 +251,18 @@ let allDemoGroups = [
 				data: bubbleDemos.bubbleDiscreteData,
 				chartType: chartTypes.BubbleChart,
 				isDemoExample: true
+			},
+			{
+				options: bubbleDemos.bubbleEmptyStateOptions,
+				data: bubbleDemos.bubbleEmptyStateData,
+				chartType: chartTypes.BubbleChart,
+				isDemoExample: true
+			},
+			{
+				options: bubbleDemos.bubbleSkeletonOptions,
+				data: bubbleDemos.bubbleSkeletonData,
+				chartType: chartTypes.BubbleChart,
+				isDemoExample: true
 			}
 		]
 	},
@@ -188,6 +272,18 @@ let allDemoGroups = [
 			{
 				options: donutDemos.donutOptions,
 				data: donutDemos.donutData,
+				chartType: chartTypes.DonutChart,
+				isDemoExample: true
+			},
+			{
+				options: donutDemos.donutEmptyStateOptions,
+				data: donutDemos.donutEmptyStateData,
+				chartType: chartTypes.DonutChart,
+				isDemoExample: true
+			},
+			{
+				options: donutDemos.donutSkeletonOptions,
+				data: donutDemos.donutSkeletonData,
 				chartType: chartTypes.DonutChart,
 				isDemoExample: true
 			}
@@ -227,6 +323,21 @@ let allDemoGroups = [
 				options: lineDemos.lineHorizontalOptions,
 				data: lineDemos.lineData,
 				chartType: chartTypes.LineChart
+			},
+			{
+				options: lineDemos.lineTimeSeriesWithThresholdsOptions,
+				data: lineDemos.lineTimeSeriesData,
+        chartType: chartTypes.LineChart
+      },
+      {
+				options: lineDemos.lineEmptyStateOptions,
+				data: lineDemos.lineEmptyStateData,
+				chartType: chartTypes.LineChart
+			},
+			{
+				options: lineDemos.lineSkeletonOptions,
+				data: lineDemos.lineSkeletonData,
+				chartType: chartTypes.LineChart
 			}
 		]
 	},
@@ -236,6 +347,18 @@ let allDemoGroups = [
 			{
 				options: pieDemos.pieOptions,
 				data: pieDemos.pieData,
+				chartType: chartTypes.PieChart,
+				isDemoExample: true
+			},
+			{
+				options: pieDemos.pieEmptyStateOptions,
+				data: pieDemos.pieEmptyStateData,
+				chartType: chartTypes.PieChart,
+				isDemoExample: true
+			},
+			{
+				options: pieDemos.pieSkeletonOptions,
+				data: pieDemos.pieSkeletonData,
 				chartType: chartTypes.PieChart,
 				isDemoExample: true
 			}
@@ -278,6 +401,18 @@ let allDemoGroups = [
 				data: scatterDemos.scatterDiscreteData,
 				chartType: chartTypes.ScatterChart,
 				isDemoExample: true
+			},
+			{
+				options: scatterDemos.scatterEmptyStateOptions,
+				data: scatterDemos.scatterEmptyStateData,
+				chartType: chartTypes.ScatterChart,
+				isDemoExample: true
+			},
+			{
+				options: scatterDemos.scatterSkeletonOptions,
+				data: scatterDemos.scatterSkeletonData,
+				chartType: chartTypes.ScatterChart,
+				isDemoExample: true
 			}
 		]
 	},
@@ -293,6 +428,18 @@ let allDemoGroups = [
 			{
 				options: stepDemos.stepTimeSeriesOptions,
 				data: stepDemos.stepTimeSeriesData,
+				chartType: chartTypes.LineChart,
+				isDemoExample: true
+			},
+			{
+				options: stepDemos.stepEmptyStateOptions,
+				data: stepDemos.stepEmptyStateData,
+				chartType: chartTypes.LineChart,
+				isDemoExample: true
+			},
+			{
+				options: stepDemos.stepSkeletonOptions,
+				data: stepDemos.stepSkeletonData,
 				chartType: chartTypes.LineChart,
 				isDemoExample: true
 			}
@@ -365,6 +512,36 @@ let allDemoGroups = [
 				data: timeSeriesAxisDemos.lineTimeSeriesNoExtendedDomainData,
 				options: timeSeriesAxisDemos.lineTimeSeriesNoExtendedDomainOptions,
 				chartType: chartTypes.LineChart
+			},
+			{
+				data: timeSeriesAxisDemos.lineTimeSeriesDataTwoIdenticalLabels,
+				options: timeSeriesAxisDemos.lineTimeSeriesTwoIdenticalLabelsOptions,
+				chartType: chartTypes.LineChart
+			},
+			{
+				data: timeSeriesAxisDemos.lineTimeSeriesDataAllLabelsInPrimaryFormat,
+				options: timeSeriesAxisDemos.lineTimeSeriesAllLabelsInPrimaryFormatOptions,
+				chartType: chartTypes.LineChart
+			}
+		]
+	},
+	{
+		title: "Radar",
+		demos: [
+			{
+				data: radarDemos.radarData,
+				options: radarDemos.radarOptions,
+				chartType: chartTypes.RadarChart
+			},
+			{
+				data: radarDemos.radarWithMissingDataData,
+				options: radarDemos.radarWithMissingDataOptions,
+				chartType: chartTypes.RadarChart
+			},
+			{
+				data: radarDemos.radarDenseData,
+				options: radarDemos.radarDenseOptions,
+				chartType: chartTypes.RadarChart
 			}
 		]
 	},