+++ packages/core/src/components/index.ts
@@ -2,6 +2,7 @@ export * from "./component";
 
 // ESSENTIALS
 export * from "./essentials/legend";
+export * from "./essentials/threshold";
 export * from "./essentials/title";
 export * from "./essentials/tooltip";
 export * from "./essentials/tooltip-bar";
@@ -19,6 +20,7 @@ export * from "./graphs/scatter";
 export * from "./graphs/pie";
 export * from "./graphs/gauge";
 export * from "./graphs/donut";
+export * from "./graphs/skeleton";
 
 // Layout
 export * from "./layout/spacer";