+++ packages/core/src/configuration.ts
@@ -144,7 +144,8 @@ const chart: BaseChartOptions = {
 		prefix: "cc"
 	},
 	data: {
-		groupMapsTo: "group"
+		groupMapsTo: "group",
+		loading: false
 	},
 	color: {
 		scale: null