+++ packages/core/stories/tutorials/getting-started/angular.ts
@@ -2,6 +2,7 @@ import marked from "marked";
 
 export const angularTutorial = marked(`
 # Carbon Charts - Angular
+
 The Carbon Charts Angular library provides a collection of reusable charting components to build websites
 and user interfaces. Adopting the library enables developers to use consistent markup, styles, and behavior
 in prototype and production work.
@@ -80,7 +81,8 @@ See tutorial on tabular data format for more information on chart data.
 
 &nbsp;
 ## Options
-You can see the options for all charts
+
+You can see the default options for all charts
 [here](https://github.com/carbon-design-system/carbon-charts/tree/master/packages/core/src/configuration.ts).
 
 &nbsp;
@@ -91,6 +93,7 @@ You can see the options for all charts
 
 &nbsp;
 ## Guidance
+
 Please refer to the
 [Carbon Design Systems guidance](https://www.carbondesignsystem.com/data-visualization/chart-types)
 on using the different charts available in this library.
@@ -119,6 +122,7 @@ View available components [here](https://github.com/carbon-design-system/carbon-
 
 &nbsp;
 ## Troubleshoot
+
 If you experience any issues while getting set up with Carbon Charts, please head over to the
 [GitHub repo](https://github.com/carbon-design-system/carbon-charts) for more guidelines and support.
 Please [create an issue](https://github.com/carbon-design-system/carbon-charts/issues) if your issue