+++ packages/core/stories/tutorials/getting-started/react.ts
@@ -2,6 +2,7 @@ import marked from "marked";
 
 export const reactTutorial = marked(`
 # Carbon Charts - React
+
 The Carbon Charts library provides a collection of reusable charting components to
 build websites and user interfaces. Adopting the library enables developers to use
 consistent markup, styles, and behavior in prototype and production work.
@@ -66,9 +67,9 @@ See tutorial on tabular data format for more information on chart data.
 
 &nbsp;
 ## Options
+
 You can see the options for all charts
 [here](https://github.com/carbon-design-system/carbon-charts/tree/master/packages/core/src/configuration.ts).
-
 
 &nbsp;
 ## Demos and Examples
@@ -78,6 +79,7 @@ You can see the options for all charts
 
 &nbsp;
 ## Guidance
+
 Please refer to the
 [Carbon Design Systems guidance](https://www.carbondesignsystem.com/data-visualization/chart-types)
 on using the different charts available in this library.
@@ -106,6 +108,7 @@ View available components [here](https://github.com/carbon-design-system/carbon-
 
 &nbsp;
 ## Troubleshoot
+
 If you experience any issues while getting set up with Carbon Charts, please head over to
 the [GitHub repo](https://github.com/carbon-design-system/carbon-charts)
 for more guidelines and support.